# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DukaanBazar Indian General Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  categoryId: number;
  categoryName: string;
  /** @nullable */
  brand?: string | null;
  stock: number;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  tags?: string[];
  createdAt: string;
}