# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DukaanBazar Indian General Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  categoryId: number;
  brand?: string;
  stock: number;
  imageUrl: string;
  images?: string[];
  isFeatured?: boolean;
  tags?: string[];
}