# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DukaanBazar Indian General Store API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  name?: string;
  phone?: string;
  address?: string;
  city?: string;
  state?: string;
  pincode?: string;
}